# src/main/resources/SwaggerConfig.yaml
swagger: "2.0"
info:
  description: "API for managing student information"
  version: "1.0.0"
  title: "Student Application API"
  termsOfService: "Free to use"
  contact:
    name: "Saiswetha"
    url: "https://editor.swagger.io/"
    email: "saiswetha0524@gmail.com"
  license:
    name: "API License"
    url: "https://editor.swagger.io/"
host: "localhost:9091"
basePath: "/"
paths:
  /studentDetails/:
    get:
      tags:
        - "Student Details"
      summary: "Get all students"
      description: "Returns all Students"
      operationId: "getAllStuds"
      responses:
        200:
          description: "Nice!"
          schema:
            $ref: "#/definitions/APIResponse"
        400:
          description: "Error happened"
          schema:
            $ref: "#/definitions/APIResponse"
        401:
          description: "Unauthorized"
        403:
          description: "Forbidden"
        404:
          description: "Student not found"
          schema:
            $ref: "#/definitions/APIResponse"
    post:
      tags:
        - "Student Details"
      summary: "Create a new student"
      description: "Adds Students"
      operationId: "additionOfStuds"
      parameters:
        - name: "student"
          in: "body"
          description: "Student object"
          required: true
          schema:
            $ref: "#/definitions/Student"
      responses:
        200:
          description: "Nice!"
          schema:
            $ref: "#/definitions/Student"
        201:
          description: "Created"
        400:
          description: "Error happened"
          schema:
            $ref: "#/definitions/APIResponse"
        401:
          description: "Unauthorized"
        403:
          description: "Forbidden"
        404:
          description: "Student cant be added/already present"
          schema:
            $ref: "#/definitions/APIResponse"
    delete:
      tags:
        - "Student Details"
      summary: "Delete every student"
      description: "Deletes All Students"
      operationId: "delAllStuds"
      responses:
        200:
          description: "Nice!"
          schema:
            $ref: "#/definitions/APIResponse"
        204:
          description: "No Content"
        400:
          description: "Error happened"
          schema:
            $ref: "#/definitions/APIResponse"
        401:
          description: "Unauthorized"
        403:
          description: "Forbidden"
        404:
          description: "Student can't be deleted"
          schema:
            $ref: "#/definitions/APIResponse"

  /studentDetails/count:
    get:
      tags:
        - "Student Details"
      summary: "Student count"
      description: "Returns Students count"
      operationId: "countOfStuds"
      responses:
        200:
          description: "Nice!"
          schema:
            type: integer
            format: int64
        400:
          description: "Error happened"
          schema:
            $ref: "#/definitions/APIResponse"
        401:
          description: "Unauthorized"
        403:
          description: "Forbidden"
        404:
          description: "Students count can't be found"
          schema:
            $ref: "#/definitions/APIResponse"

  /studentDetails/hi:
    get:
      tags:
        - "Welcome"
      summary: "Welcome message"
      description: "Hello method"
      operationId: "welcomeAll"
      responses:
        200:
          description: "Nice!"
          schema:
            type: string
        401:
          description: "Unauthorized"
        403:
          description: "Forbidden"
        404:
          description: "Not Found"

  /studentDetails/{id}:
    get:
      tags:
        - "Student Details"
      summary: "Find Student with id"
      description: "Returns Students with id"
      operationId: "getStudWithId"
      parameters:
        - in: "path"
          name: "id"
          description: "id"
          required: true
          type: "integer"
          format: "int32"
      responses:
        200:
          description: "Nice!"
          schema:
            $ref: "#/definitions/APIResponse"
        400:
          description: "Error happened"
          schema:
            $ref: "#/definitions/APIResponse"
        401:
          description: "Unauthorized"
        403:
          description: "Forbidden"
        404:
          description: "Students not found"
          schema:
            $ref: "#/definitions/APIResponse"
    delete:
      tags:
        - "Student Details"
      summary: "Delete with id"
      description: "Deletes student with id"
      operationId: "delStudWithId"
      parameters:
        - in: "path"
          name: "id"
          description: "id"
          required: true
          type: "integer"
          format: "int32"
      responses:
        200:
          description: "Nice!"
          schema:
            $ref: "#/definitions/APIResponse"
        204:
          description: "No content"
        400:
          description: "Error happened"
          schema:
            $ref: "#/definitions/APIResponse"
        401:
          description: "Unauthorized"
        403:
          description: "Forbidden"
        404:
          description: "Students can't be deleted"
          schema:
            $ref: "#/definitions/APIResponse"
definitions:
  Student:
    type: object
    properties:
      id:
        type: integer
        format: int32
      name:
        type: string


  APIResponse:
    type: object
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string